TYRUS-234: Removed support for deprecated protocol versions.
- lots of related refactorings, some renames
- removed big part of unused code
- tested on standalone Grizzly and integrated into Glassfish trunk
