+++ protocol/websocket/src/main/java/org/glassfish/tyrus/websockets/SecKey.java
@@ -51,30 +51,27 @@ import java.util.Random;
  *
  * @author Alexey Stashok
  */
-public class SecKey {
+class SecKey {
     private static final Random random = new SecureRandom();
 
-    public static final int KEY_SIZE = 16;
+    private static final int KEY_SIZE = 16;
 
     /**
      * Security key string representation, which includes chars and spaces.
      */
     private final String secKey;
-    private String nonce;
-
-    private byte[] bytes;
 
     public SecKey() {
         secKey = create();
     }
 
     private String create() {
-        bytes = new byte[KEY_SIZE];
+        byte[] bytes = new byte[KEY_SIZE];
         random.nextBytes(bytes);
         return Base64Utils.encodeToString(bytes, false);
     }
 
-    SecKey(String base64) {
+    public SecKey(String base64) {
         if (base64 == null) {
             throw new HandshakeException("Null keys are not allowed.");
         }
@@ -119,20 +116,6 @@ public class SecKey {
     @Override
     public String toString() {
         return secKey;
-    }
-
-    public byte[] getBytes() {
-        if (bytes == null) {
-            bytes = Base64Utils.decode(secKey);
-        }
-        return bytes;
-    }
-
-    public String getNonce() {
-        if (nonce == null) {
-            nonce = create();
-        }
-        return nonce;
     }
 
     public void validateServerKey(String serverKey) {
