+++ protocol/websocket/src/main/java/org/glassfish/tyrus/websockets/frame/ClosingFrame.java
@@ -1,7 +1,7 @@
 /*
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  *
- * Copyright (c) 2012 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012-2013 Oracle and/or its affiliates. All rights reserved.
  *
  * The contents of this file are subject to the terms of either the GNU
  * General Public License Version 2 only ("GPL") or the Common Development
@@ -38,34 +38,21 @@
  * holder.
  */
 
-package org.glassfish.tyrus.websockets.frametypes;
+package org.glassfish.tyrus.websockets.frame;
 
-import org.glassfish.tyrus.websockets.BaseFrameType;
 import org.glassfish.tyrus.websockets.DataFrame;
-import org.glassfish.tyrus.websockets.StrictUtf8;
-import org.glassfish.tyrus.websockets.Utf8Utils;
 import org.glassfish.tyrus.websockets.WebSocket;
 
-public class TextFrameType extends BaseFrameType {
+public class ClosingFrame extends BaseFrame {
+
     @Override
-    public void setPayload(DataFrame frame, byte[] data) {
-        frame.setPayload(data);
+    public DataFrame create(boolean fin, byte[] data) {
+        return new org.glassfish.tyrus.websockets.ClosingFrame(data);
     }
 
     @Override
-    public byte[] getBytes(DataFrame dataFrame) {
-        final byte[] bytes = dataFrame.getBytes();
-        if (bytes == null) {
-            setPayload(dataFrame, Utf8Utils.encode(new StrictUtf8(), dataFrame.getTextPayload()));
-        }
-        return dataFrame.getBytes();
-    }
-
     public void respond(WebSocket socket, DataFrame frame) {
-        if (frame.isLast()) {
-            socket.onMessage(frame.getTextPayload());
-        } else {
-            socket.onFragment(frame.isLast(), frame.getTextPayload());
-        }
+        socket.onClose((org.glassfish.tyrus.websockets.ClosingFrame) frame);
+        socket.close();
     }
 }
