+++ protocol/websocket/src/main/java/org/glassfish/tyrus/websockets/uri/internal/AbstractMultivaluedMap.java
@@ -74,7 +74,7 @@ public abstract class AbstractMultivaluedMap<K, V> implements MultivaluedMap<K, 
     /**
      * Set the value for the key to be a one item list consisting of the supplied
      * value. Any existing values will be replaced.
-     * <p />
+     * <p/>
      * NOTE: This implementation ignores {@code null} values; A supplied value
      * of {@code null} is ignored and not added to the purged value list.
      * As a result of such operation, empty value list would  be registered for
@@ -99,7 +99,7 @@ public abstract class AbstractMultivaluedMap<K, V> implements MultivaluedMap<K, 
 
     /**
      * Define the behavior for adding a {@code null} values to the value list.
-     * <p />
+     * <p/>
      * Default implementation is a no-op, i.e. the {@code null} values are ignored.
      * Overriding implementations may modify this behavior by providing their
      * own definitions of this method.
@@ -115,7 +115,7 @@ public abstract class AbstractMultivaluedMap<K, V> implements MultivaluedMap<K, 
     /**
      * Define the behavior for adding a {@code null} values to the first position
      * in the value list.
-     * <p />
+     * <p/>
      * Default implementation is a no-op, i.e. the {@code null} values are ignored.
      * Overriding implementations may modify this behavior by providing their
      * own definitions of this method.
@@ -130,7 +130,7 @@ public abstract class AbstractMultivaluedMap<K, V> implements MultivaluedMap<K, 
 
     /**
      * Add a value to the current list of values for the supplied key.
-     * <p />
+     * <p/>
      * NOTE: This implementation ignores {@code null} values; A supplied value
      * of {@code null} is ignored and not added to the value list. Overriding
      * implementations may modify this behavior by redefining the
@@ -155,7 +155,7 @@ public abstract class AbstractMultivaluedMap<K, V> implements MultivaluedMap<K, 
      * the supplied array of new values is empty, method returns immediately.
      * Method throws a {@code NullPointerException} if the supplied array of values
      * is {@code null}.
-     * <p />
+     * <p/>
      * NOTE: This implementation ignores {@code null} values; Any of the supplied values
      * of {@code null} is ignored and not added to the value list. Overriding
      * implementations may modify this behavior by redefining the
@@ -190,7 +190,7 @@ public abstract class AbstractMultivaluedMap<K, V> implements MultivaluedMap<K, 
      * values for the supplied key. If the supplied value list is empty, method
      * returns immediately. Method throws a {@code NullPointerException} if the
      * supplied array of values is {@code null}.
-     * <p />
+     * <p/>
      * NOTE: This implementation ignores {@code null} values; Any {@code null} value
      * in the supplied value list is ignored and not added to the value list. Overriding
      * implementations may modify this behavior by redefining the
@@ -233,7 +233,7 @@ public abstract class AbstractMultivaluedMap<K, V> implements MultivaluedMap<K, 
     /**
      * Add a value to the first position in the current list of values for the
      * supplied key.
-     * <p />
+     * <p/>
      * NOTE: This implementation ignores {@code null} values; A supplied value
      * of {@code null} is ignored and not added to the purged value list. Overriding
      * implementations may modify this behavior by redefining the
@@ -256,7 +256,7 @@ public abstract class AbstractMultivaluedMap<K, V> implements MultivaluedMap<K, 
     /**
      * Return a non-null list of values for a given key. The returned list may be
      * empty.
-     * <p />
+     * <p/>
      * If there is no entry for the key in the map, a new empty {@link List}
      * instance is created, registered within the map to hold the values of
      * the key and returned from the method.
@@ -281,7 +281,7 @@ public abstract class AbstractMultivaluedMap<K, V> implements MultivaluedMap<K, 
 
     /**
      * {@inheritDoc }
-     * <p />
+     * <p/>
      * This implementation delegates the method call to to the the underlying
      * [key, multi-value] store.
      *
@@ -294,7 +294,7 @@ public abstract class AbstractMultivaluedMap<K, V> implements MultivaluedMap<K, 
 
     /**
      * {@inheritDoc }
-     * <p />
+     * <p/>
      * This implementation delegates the method call to to the the underlying
      * [key, multi-value] store.
      *
