+++ protocol/websocket/src/main/java/org/glassfish/tyrus/websockets/uri/internal/MultivaluedHashMap.java
@@ -48,7 +48,7 @@ import java.util.Map;
 
 /**
  * A hash table based implementation of {@link MultivaluedMap} interface.
- *
+ * <p/>
  * This implementation provides all of the optional map operations. This class
  * makes no guarantees as to the order of the map; in particular, it does not
  * guarantee that the order will remain constant over time. The implementation
@@ -56,7 +56,7 @@ import java.util.Map;
  * {@code null} values, but ignores them. This behavior can be customized
  * by overriding the protected {@link #addNull(List) addNull(...)} and
  * {@link #addFirstNull(List) addFirstNull(...)} methods.
- * <p />
+ * <p/>
  * This implementation provides constant-time performance for the basic
  * operations (<tt>get</tt> and <tt>put</tt>), assuming the hash function
  * disperses the elements properly among the buckets. Iteration over
@@ -65,7 +65,7 @@ import java.util.Map;
  * of key-value mappings).  Thus, it's very important not to set the initial
  * capacity too high (or the load factor too low) if iteration performance is
  * important.
- * <p />
+ * <p/>
  * An instance of <tt>MultivaluedHashMap</tt> has two parameters that affect its
  * performance: <i>initial capacity</i> and <i>load factor</i>. The <i>capacity</i>
  * is the number of buckets in the hash table, and the initial capacity is simply
@@ -75,7 +75,7 @@ import java.util.Map;
  * the product of the load factor and the current capacity, the hash table is
  * <i>rehashed</i> (that is, internal data structures are rebuilt) so that the
  * hash table has approximately twice the number of buckets.
- * <p />
+ * <p/>
  * As a general rule, the default load factor (.75) offers a good tradeoff
  * between time and space costs. Higher values decrease the space overhead
  * but increase the lookup cost (reflected in most of the operations of the
@@ -85,12 +85,12 @@ import java.util.Map;
  * number of rehash operations. If the initial capacity is greater
  * than the maximum number of entries divided by the load factor, no
  * rehash operations will ever occur.
- * <p />
+ * <p/>
  * If many mappings are to be stored in a <tt>MultivaluedHashMap</tt> instance,
  * creating it with a sufficiently large capacity will allow the mappings to
  * be stored more efficiently than letting it perform automatic rehashing as
  * needed to grow the table.
- * <p />
+ * <p/>
  * <strong>Note that this implementation is not guaranteed to be synchronized.</strong>
  * If multiple threads access a hash map concurrently, and at least one of
  * the threads modifies the map structurally, it <i>must</i> be
@@ -99,7 +99,7 @@ import java.util.Map;
  * associated with a key that an instance already contains is not a
  * structural modification.) This is typically accomplished by
  * synchronizing on some object that naturally encapsulates the map.
- * <p />
+ * <p/>
  * The iterators returned by all of this class's "collection view methods"
  * are <i>fail-fast</i>: if the map is structurally modified at any time after
  * the iterator is created, in any way except through the iterator's own
@@ -107,7 +107,7 @@ import java.util.Map;
  * Thus, in the face of concurrent modification, the iterator fails quickly and
  * cleanly, rather than risking arbitrary, non-deterministic behavior at an
  * undetermined time in the future.
- * <p />
+ * <p/>
  * Note that the fail-fast behavior of an iterator cannot be guaranteed
  * as it is, generally speaking, impossible to make any hard guarantees in the
  * presence of unsynchronized concurrent modification. Fail-fast iterators
