+++ protocol/websocket/src/main/java/org/glassfish/tyrus/websockets/uri/internal/PathSegment.java
@@ -57,7 +57,7 @@ public interface PathSegment {
 
     /**
      * Get the path segment.
-     * <p>
+     * <p/>
      *
      * @return the path segment
      */
