+++ protocol/websocket/src/main/java/org/glassfish/tyrus/websockets/uri/internal/UriComponent.java
@@ -423,17 +423,17 @@ public class UriComponent {
     /**
      * Decodes characters of a string that are percent-encoded octets using
      * UTF-8 decoding (if needed).
-     * <p>
+     * <p/>
      * It is assumed that the string is valid according to an (unspecified) URI
      * component type. If a sequence of contiguous percent-encoded octets is
      * not a valid UTF-8 character then the octets are replaced with '\uFFFD'.
-     * <p>
+     * <p/>
      * If the URI component is of type HOST then any "%" found between "[]" is
      * left alone. It is an IPv6 literal with a scope_id.
-     * <p>
+     * <p/>
      * If the URI component is of type QUERY_PARAM then any "+" is decoded as
      * as ' '.
-     * <p>
+     * <p/>
      *
      * @param s the string to be decoded.
      * @param t the URI component type, may be null.
@@ -601,7 +601,7 @@ public class UriComponent {
 
     /**
      * Decode the path component of a URI as path segments.
-     * <p>
+     * <p/>
      * Any '/' character in the path is considered to be a deliminator
      * between two path segments. Thus if the path is '/' then the path segment
      * list will contain two empty path segments. If the path is "//" then
@@ -767,7 +767,7 @@ public class UriComponent {
 
     /**
      * Decode a contigious sequence of percent encoded octets.
-     * <p>
+     * <p/>
      * Assumes the index, i, starts that the first hex digit of the first
      * percent-encoded octet.
      */
@@ -861,6 +861,7 @@ public class UriComponent {
 
     /**
      * Checks whether the character {@code c} is hexadecimal character.
+     *
      * @param c Any character
      * @return The is {@code c} is a hexadecimal character (e.g. 0, 5, a, A, f, ...)
      */
