+++ protocol/websocket/src/main/java/org/glassfish/tyrus/websockets/uri/internal/l10n/Localizable.java
@@ -49,8 +49,8 @@ public interface Localizable {
     /**
      * Special constant that represents a message that
      * is not localizable.
-     *
-     * <p>
+     * <p/>
+     * <p/>
      * Use of "new" is to create an unique instance.
      */
     public static final String NOT_LOCALIZABLE = "\u0000";
@@ -59,8 +59,8 @@ public interface Localizable {
      * Gets the key in the resource bundle.
      *
      * @return if this method returns {@link #NOT_LOCALIZABLE}, that means the
-     *     message is not localizable, and the first item of {@link #getArguments()}
-     *     array holds a {@code String}.
+     *         message is not localizable, and the first item of {@link #getArguments()}
+     *         array holds a {@code String}.
      */
     public String getKey();
